dynamic util list

+++ scripts/run-tests.sh
@@ -9,6 +9,13 @@ cargo build --release
 
 ./scripts/gen-test-helper.sh
 
+mapfile -t UTILS < "$PROJECT_DIR/.test-helpers/utils.list"
+
+if [[ ${#UTILS[@]} -eq 0 ]]; then
+    echo "ERROR: .test-helpers/utils.list is empty; nothing to test." >&2
+    exit 1
+fi
+
 # Clear stale diff files from any previous run
 rm -rf "$PROJECT_DIR/.test-helpers/tests/diff"
 
@@ -17,7 +24,7 @@ set +e
 "$GNU_PROJECT_DIR/tests/run.sh" \
     --builddir="$PROJECT_DIR/.test-helpers" \
     "$@" \
-    cal dmesg hexdump lscpu lslocks lsmem
+    "${UTILS[@]}"
 GNU_TEST_EXIT=$?
 set -e
